test: respect Fireworks contract test boundaries
